+++ config.py
@@ -1,0 +1,2 @@
+SIZE = 40
+BACKGROUND_COLOR = (110, 110, 5)
